korg-volca-trs-midi: v2! don't connect DIN jack shield to GND
MIDI cables shield the signal with ground, but that doesn't mean the
sleeve of the DIN jack is also connected to GND. The previous revision
connected them, and I suspect that's the culprit of terrible noise
interference when connected.

(I confirmed the original Volca MIDI DIN jack sleeve/shield is not
connected to MIDI GND, nor to the battery negative terminal, which may
or may not mean anything).